Setup editor: preset provenance, panel chrome, fractional blinds; calendar week-chip fix
Blind editor
- +Round / +Break removed from the rail: the row menu inserts relative to
  where you are, which a rail button cannot. An empty grid gets an empty
  state (first round / break / generator) since there is no row to
  right-click — without it, deleting every row was a dead end.
- Blinds are money, not just chips. The classic timer has always stored them
  as floats; the grid's (int) casts were a regression that turned 0.25 into 0
  on every copy-on-write save. Money cells parse as floats to 2dp with
  step="any" and inputmode="decimal" (a numeric keypad has no decimal point).
  The ladder extends below 1, but rungs >= 1 stay whole — half a chip does
  not exist, and rounding the whole ladder changed every generated structure.
  fmtChips() in timer_beta.js gained the fraction-digit branch timer.js had,
  or a .50 blind rendered as "0.5".
- Opening the generator scrolls it into view and labels its state; it opens
  below a 20-row schedule, so the button read as doing nothing.
- Long-press works over a cell, matching a desktop right-click. The field is
  unselectable only for the press duration — a blanket user-select:none on an
  input risks taking the caret with it on older iOS.

Setup editor
- The panel is a layer over the console now, not another white card in it:
  dark chrome header, toned tab strip, elevation, and a gap above.
- Game Preset is state-aware: "From: X", "From: X • MODIFIED", or "Not from a
  preset". Provenance is stored on the session (preset_structure_id) rather
  than a JS variable that a reload forgot, which is why the bar could only
  ever say "Custom (unsaved)". Drift is computed server-side against the
  preset AS IT EXISTS NOW, not a fingerprint taken at load time, and only
  across sections the preset stores so legacy rows are not permanently
  modified. "Update preset" writes changes back; Delete / Set default moved
  into a ⋯ menu. save_payout_structure required a name, so every update was
  rejected before reaching the update branch.
- Blinds and Timer tabs follow the Game Type dropdown immediately in both
  directions; they were built only for a saved tournament, so switching to
  Cash left them standing and switching to Tournament could not reveal them.
  They stay disabled until saved because event_setup_dl refuses a session
  that is not already a saved tournament.
- The check-in header parks below the sticky site nav instead of behind it.

Calendar
- Week-view chips passed the whole event object to viewEvent(), which
  concatenates its first argument into the URL: every chip went to
  /event.php?id=[object Object]. Month view was fixed in v0.2083; these two
  JS callers were missed. Chips now also pass the cell's own date, so a
  multi-day event opens the day you clicked rather than day one.

Docs
- SECURITY.md gains sweep 1a: parse errors in PHP-generated JavaScript.
  php -l cannot see them, and one SyntaxError discards the whole <script>
  block, so the page renders and nothing works. Cost the check-in console
  during this work.

+++ www/event_blinds.php
@@ -102,8 +102,9 @@ $csrf = csrf_token();
 var ES_CSRF = <?= json_encode($csrf) ?>;
 var ES_EVENT_ID = <?= (int)$event_id ?>;
 var ES_LEVELS = <?= json_encode(array_map(function ($l) {
-    return ['small_blind' => (int)$l['small_blind'], 'big_blind' => (int)$l['big_blind'],
-            'ante' => (int)$l['ante'], 'duration_minutes' => (int)$l['duration_minutes'],
+    // (float) for the money columns — see pk_clean_blind_levels().
+    return ['small_blind' => (float)$l['small_blind'], 'big_blind' => (float)$l['big_blind'],
+            'ante' => (float)$l['ante'], 'duration_minutes' => (int)$l['duration_minutes'],
             'is_break' => (int)$l['is_break']];
 }, $levels)) ?>;
 var ES_CURRENT_LEVEL = <?= (int)$current_level ?>;
